--- assignments/01-programming/02-control-flow/09-collatz/submissions/joaomanoelfmn/src/CollatzSequence.java
@@ -1,49 +1,0 @@
-import java.util.Scanner;
-
-public class CollatzSequence {
-
-    public static int nextCollatz(int n) {
-        if (n % 2 == 0) {
-            return n / 2;
-        }
-        else {
-            return (n * 3) + 1;
-        }
-    }
-
-    public static int calculateCollatzSum(int n) {
-        int sum = n;
-
-        while (n != 1) {
-            n = nextCollatz(n);
-            sum += n;
-        }
-        
-        return sum;
-    }
-
-    public static void main(String[] args) {
-        Scanner scanner = new Scanner(System.in);
-
-        System.out.print("Digite um número inteiro positivo: ");
-        int n = scanner.nextInt();
-
-        if (n <= 0) {
-            System.out.println("Erro: O número deve ser um inteiro positivo.");
-        } else {
-            System.out.print("Sequência de Collatz: " + n);
-            
-            int current = n; 
-            while (current != 1) {
-                current = nextCollatz(current);
-                System.out.print(" → " + current);
-            }
-            System.out.println();
-
-            int totalSum = calculateCollatzSum(n);
-            System.out.println("Soma dos termos: " + totalSum);
-        }
-
-        scanner.close();
-    }
-}
